padding-bottom added to main-part
want to see how it's going to look on mobile with a padding-bottom added to a main-part (part with quote), check if this make a composition looks still good. If not, keep working on that. If yes, style a button to look better, add hover, and write rules for position for other types of screen.